+++ kroxylicious-runtime/src/test/java/io/kroxylicious/proxy/config/RoutingGraphWalkerTest.java
@@ -19,147 +19,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 class RoutingGraphWalkerTest {
 
     // -------------------------------------------------------------------------
-    // anyInRouterGraph
-    // -------------------------------------------------------------------------
-
-    @Test
-    void returnsFalseWhenNeitherPredicateMatches() {
-        // Given
-        var routers = Map.of("r1", routerDef("r1", "cluster-a"));
-        var clusters = Map.of("cluster-a", clusterDef("cluster-a"));
-
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInRouterGraph("r1", routers, clusters, name -> false, name -> false)).isFalse();
-    }
-
-    @Test
-    void returnsTrueWhenRouterNamePredicateMatches() {
-        // Given
-        var routers = Map.of("r1", routerDef("r1", "cluster-a"));
-
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInRouterGraph("r1", routers, Map.of(), "r1"::equals, name -> false)).isTrue();
-    }
-
-    @Test
-    void returnsTrueWhenClusterLeafPredicateMatches() {
-        // Given
-        var routers = Map.of("r1", routerDef("r1", "cluster-a"));
-        var clusters = Map.of("cluster-a", clusterDef("cluster-a"));
-
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInRouterGraph("r1", routers, clusters, name -> false, "cluster-a"::equals)).isTrue();
-    }
-
-    @Test
-    void returnsFalseWhenClusterLeafDoesNotMatchPredicate() {
-        // Given
-        var routers = Map.of("r1", routerDef("r1", "cluster-a"));
-        var clusters = Map.of("cluster-a", clusterDef("cluster-a"));
-
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInRouterGraph("r1", routers, clusters, name -> false, "cluster-b"::equals)).isFalse();
-    }
-
-    @Test
-    void walksThroughNestedRouters() {
-        // Given
-        var r2 = routerDef("r2", "cluster-deep");
-        var r1 = routerDefWithRouterTarget("r1", "r2");
-        var routers = Map.of("r1", r1, "r2", r2);
-        var clusters = Map.of("cluster-deep", clusterDef("cluster-deep"));
-
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInRouterGraph("r1", routers, clusters, name -> false, "cluster-deep"::equals)).isTrue();
-    }
-
-    @Test
-    void returnsFalseForUnknownEntryPointRouter() {
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInRouterGraph("unknown", Map.of(), Map.of(), name -> false, name -> false)).isFalse();
-    }
-
-    @Test
-    void doesNotLoopOnCycle() {
-        // Given
-        var r1 = routerDefWithRouterTarget("r1", "r2");
-        var r2 = routerDefWithRouterTarget("r2", "r1");
-        var routers = Map.of("r1", r1, "r2", r2);
-
-        // When
-        // Then
-        assertThat(RoutingGraphWalker.anyInRouterGraph("r1", routers, Map.of(), name -> false, name -> false)).isFalse();
-    }
-
-    @Test
-    void returnsTrueWhenNestedRouterNameMatchesPredicate() {
-        // Given
-        var r2 = routerDef("r2", "cluster-a");
-        var r1 = routerDefWithRouterTarget("r1", "r2");
-        var routers = Map.of("r1", r1, "r2", r2);
-
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInRouterGraph("r1", routers, Map.of(), "r2"::equals, name -> false)).isTrue();
-    }
-
-    // -------------------------------------------------------------------------
-    // anyInClusterGraph
-    // -------------------------------------------------------------------------
-
-    @Test
-    void vcWithNamedClusterReturnsTrueWhenClusterPredicateMatches() {
-        // Given
-        var vc = vcWithNamedCluster("vc1", "cluster-a");
-        var clusters = Map.of("cluster-a", clusterDef("cluster-a"));
-
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInClusterGraph(vc, Map.of(), clusters, name -> false, "cluster-a"::equals)).isTrue();
-    }
-
-    @Test
-    void vcWithNamedClusterReturnsFalseWhenClusterPredicateDoesNotMatch() {
-        // Given
-        var vc = vcWithNamedCluster("vc1", "cluster-a");
-        var clusters = Map.of("cluster-a", clusterDef("cluster-a"));
-
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInClusterGraph(vc, Map.of(), clusters, name -> false, "cluster-b"::equals)).isFalse();
-    }
-
-    @Test
-    void vcWithRouterReturnsTrueWhenRouterGraphContainsMatchingCluster() {
-        // Given
-        var routers = Map.of("r1", routerDef("r1", "cluster-a"));
-        var clusters = Map.of("cluster-a", clusterDef("cluster-a"));
-        var vc = vcWithRouter("vc1", "r1");
-
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInClusterGraph(vc, routers, clusters, name -> false, "cluster-a"::equals)).isTrue();
-    }
-
-    @Test
-    void vcWithRouterReturnsFalseWhenRouterGraphContainsNoMatchingCluster() {
-        // Given
-        var routers = Map.of("r1", routerDef("r1", "cluster-a"));
-        var clusters = Map.of("cluster-a", clusterDef("cluster-a"));
-        var vc = vcWithRouter("vc1", "r1");
-
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInClusterGraph(vc, routers, clusters, name -> false, "cluster-b"::equals)).isFalse();
-    }
-
-    @Test
-    @SuppressWarnings("deprecation")
-    void vcWithInlineTargetClusterReturnsFalse() {
-        // Given
-        var vc = new VirtualCluster("vc1", new TargetCluster("kafka:9092", Optional.empty()),
-                List.of(gateway()), false, false, null);
-
-        // When / Then
-        assertThat(RoutingGraphWalker.anyInClusterGraph(vc, Map.of(), Map.of(), name -> true, name -> true)).isFalse();
-    }
-
-    // -------------------------------------------------------------------------
     // walkClusterGraph — visitor is called for all node types
     // -------------------------------------------------------------------------
